+++ compiler/src/test/kotlin/software/amazon/test/Singleton.kt
@@ -4,4 +4,4 @@ import me.tatarka.inject.annotations.Scope
 
 @Scope
 @Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY_GETTER)
-annotation class SingleInAppScope
+annotation class Singleton